[#259] refactor: 수정요청사항 설명 공통 컴포넌트로 분리
